Partial revert prior, suppress FileLength, add doc
