+++ config/checkstyle/suppressions.xml
@@ -78,6 +78,7 @@
     <!--DBCollection is insanely long, and we should not compromise every class length for this one-->
     <suppress checks="FileLength" files="DBCollection"/>
     <suppress checks="FileLength" files="MongoCollection"/>
+    <suppress checks="FileLength" files="Aggregates"/>
 
     <suppress checks="ParameterNumber" files="Connection"/>
     <suppress checks="ParameterNumber" files="DefaultServer"/>
