+++ driver-core/src/main/com/mongodb/client/model/Aggregates.java
@@ -1007,6 +1007,53 @@ public final class Aggregates {
         }
     }
 
+    private static class SimplePipelineStage implements Bson {
+        private final String name;
+        private final Bson value;
+
+        SimplePipelineStage(final String name, final Bson value) {
+            this.name = name;
+            this.value = value;
+        }
+
+        @Override
+        public <TDocument> BsonDocument toBsonDocument(final Class<TDocument> documentClass, final CodecRegistry codecRegistry) {
+            return new BsonDocument(name, value.toBsonDocument(documentClass, codecRegistry));
+        }
+
+        @Override
+        public boolean equals(final Object o) {
+            if (this == o) {
+                return true;
+            }
+            if (o == null || getClass() != o.getClass()) {
+                return false;
+            }
+
+            SimplePipelineStage that = (SimplePipelineStage) o;
+
+            if (name != null ? !name.equals(that.name) : that.name != null) {
+                return false;
+            }
+            return value != null ? value.equals(that.value) : that.value == null;
+        }
+
+        @Override
+        public int hashCode() {
+            int result = name != null ? name.hashCode() : 0;
+            result = 31 * result + (value != null ? value.hashCode() : 0);
+            return result;
+        }
+
+        @Override
+        public String toString() {
+            return "Stage{"
+                           + "name='" + name + '\''
+                           + ", value=" + value
+                           + '}';
+        }
+    }
+
     private static final class BucketStage<TExpression, TBoundary> implements Bson {
 
         private final TExpression groupBy;
